+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed Serializer to SceneSerializer
 - Instead of having one class that handles serialization of all components, the design was changed to have individual ComponentSerializer structs that can be implemented via template specialization.
 - The same thing is done for the GUI rendering of components
+- Moved SceneHierarchyPanel and PerformancePanel to GUI namespace
 
 ### Fixed